Refactored Auth Services
