+++ lib/src/features/authentication/presentation/pages/registration_page.dart
@@ -1,4 +1,4 @@
-import 'package:apk_iman_ba/Services/auth_service.dart';
+import 'package:apk_iman_ba/src/features/authentication/data/repositories/auth_repository.dart';
 import 'package:apk_iman_ba/src/shared/common_widgets/custom_textfield.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
